Did refactoring. Implemented class Request
